Fix install failure when using non-version tag
When installing a package with a non-version formatted git tag
(for example 'example_tag') as the as the most recent tag then
setuptools_scm will assert that it can't parse the tag.  This patch
fixes this assert by filtering out non-version formatted tags.  By
default any tag with a '.' in the name will be interpreted as a version
tag.  Optionally, if 'exact' is set to true when calling parse then
only tags in the form of 'v*.*.*' will be found.