properly write version file even if in pretend mode
before this, SETUPTOOLS_SCM_PRETEND_VERSION make setuptools_scm fail
to write the `write_to` file correctly. this was apparently to avoid
issues with archives, which we believe is not a problem anymore.

this *is*, however, a problem for other tools like Debian packaging:
there, we use the pretend mechanism to import the version number from
the Debian packaging system (from `debian/changelog`, usually). this
works at setup time: setuptools_scm has the proper environment and
detects the version. but if the software was written to depend on the
`write_to` file, it will fail to find that version at runtime because
`setup.py` will not have written the `write_to` file properly at build
time.

this patch simply makes sure that the version is written even if the
provided version number is a string, which is usually the case for
pretend version numbers. a fixed test case is also provided.

Closes: #101