Prefer to only use 'git describe' for all required info
Previously, three git commands were used to gather all required info.
Now all info is gathered from a single 'git describe' call in all usual
cases, and only if that fails fallbacks are used.

The --dirty option was already present before commit 929a5d7, but it was
removed to address [1], which turns out to not actually be necessary as
we can try "git describe --dirty" first and then eventually use the
fallback.

[1] https://github.com/pypa/setuptools_scm/issues/86