file finder: symlink edge cases
It turns out we had a edge case to handle, as the algorithm
included scm controlled files reachable through symlinks while the
symlink itself was not scm controlled.

Additionally, better handle situations where we include a directory
below 'path' as well as a symlink to it. Before it would randomly
return the original directory, or include it under the symlink.