Allow getting the version from the parent directory suffix.
Add a `parentdir_prefix_version` config parameter which allows lets
setuptools_scm get a version string from the name of the parent
directory: if `parentdir_prefix_version = "foo-"` and the parent
directory is named `foo-v12.34`, then the version is 12.34.  This is
only active if the parameter is set (it defaults to unset, i.e. the
feature is opt-in), if getting the version from the SCM failed, *and*
there's no metadata already present.  However, this has higher priority
than `fallback_version` (which remains the last-resort fallback).

This feature is intended to support the tarballs that GitHub
automatically creates for all releases, which have no metadata but
follow the correct naming scheme for the feature to work.

This feature is directly inspired from versioneer's `parentdir_prefix`
setting.