Select git tags based on presence of numbers, rather than dots.
The default `git describe` command had an undocumented `--match *.*` in it,
intended to skip over tags that may represent human-readable checkpoints.

Unfortunately this means it would be ignoring single-component versions.

Instead look for presence of digits in the tag, it's still a heuristic, but
it should cover all previous cases, and fix single component versions.