Allow guess_next_simple_semver to ignore unneeded parts
guess_next_simple_semver previously would error on trying to convert the
'rc0' part of a version such as '1.2.0rc1' to an int, even if it was
told to retain only the first two parts. (It would actually only error
on Python 2, where the conversion to int was non-lazy)

Now if it's told to retain only the first two parts, it won't try to
convert anything after the first two dots to an int, so it won't error.
The outcome is the same as if the bit after the second dot was a valid
int (which was going to be ignored anyway, we just now don't error on
it).

This is useful for the release-branch-semver scheme, which wants to be
able to find the next minor version even if the most recent version is a
release candidate.