Consider configuration when running `__main__`
Since the usage with `setup.py` is deprecated, it would be nice to have
a command that give the same results as:

    python setup.py --version

The idea of the changes implemented here is to do that using the
`__main__` module.

The methodology consists in using the `find_files` function to look for
a `pyproject.toml` file and then reading the configuration from there.

If `pyproject.toml` file is not found or it does not contain a
`[tool.setuptools_scm]` table, the command works as implemented before
(guessing the next version).