fix: empty environment variable breaks versioning
`pretended` should only have an effect if it is set to something - anything, even `"0"`, but not if it's an empty string. This restores that behavior instead of setting 0.0.0 automatically.

If `SETUPTOOLS_SCM_PRETEND_VERSION_FOR_X` is an empty string, this will also cause the native behavior to be restored, giving a way to unset the PRETEND_VERSION for a single package.