Fix parsing .git_archival.txt for archives made from tagged commits.
Given .git_archival.txt content like this:

    node: 3634907645428b9542cfb4fd1ceef08f14526cb8
    node-date: 2022-10-13T20:38:09-06:00
    describe-name: v0.0.5
    ref-names: HEAD -> main, tag: v0.0.5

we would previously fail with an exception like this:

    $ python3 -m setuptools_scm
    Traceback (most recent call last):
      [..snip..]
      File ".../setuptools_scm/git.py", line 312, in parse_archival
        return archival_to_version(data, config=config)
      File ".../setuptools_scm/git.py", line 286, in archival_to_version
        tag, number, node, _ = _git_parse_describe(archival_describe)
      File ".../setuptools_scm/git.py", line 245, in _git_parse_describe
        tag, number, node = describe_output.rsplit("-", 2)
    ValueError: not enough values to unpack (expected 3, got 1)

The problem is that the describe-name field is "v0.0.5" instead of
looking like "v0.0.5-0-g3634907".

When we invoke git-describe ourselves, we specify the --long option, but
the format-strings implemented by git-archive do not currently give us
any way to force long-output in a .git_archival.txt file.

This change detects this condition and returns None for the missing
fields.